Replace *Tox as a pointer to *Messenger to it's own struct
Oh, boy... I knew this commit was going to be huge, but I didn't know how big.
There's no way to make this smaller, without breaking more stuff, as it is, it
only compiles if you
./configure --disable-tests --disable-testing --disable-av ...

For the most of it, It's a lot of sane renames, and close consideration. With
the notable exception of group.{c,h} & tox_old_code.h. Both of which I decided
to do bulk replacements of *Group_Chats to *Tox, and linking the code on line 1
of the function to avoid changing a lot of lines with important code. I decided
to do it this way because the global plan for toxcore is to replace the current
groupchats with something better. So this change was quick, instead of detailed.