Make file connection stateless
Messenger is slightly twisty when it comes to sending connection status
callbacks It will very likely need at the very least a partial refactor to
clean it up a bit. Toxcore shouldn't need void *userdata as deep as is
currently does.

(amend 1) Because of the nature of toxcore connection callbacks, I also decided
to include the stateless change for friend requests. It's simpler this was and
doesn't include doing anything foolish in the meantime between commits.