ping long-lived onion announce nodes and paths less often
If an announce node and its path have been around for a while and are
not showing signs of failure, they are considered 'stable' and the node
is pinged much less often. We still make at least one ping to some
announce node per ONION_NODE_PING_INTERVAL. The timeout before the
network is considered down is correspondingly increased.