remove temporary group connections when no longer needed
(even in the case that a connection is acting as an introducer connection in
both directions.)

The "peer kill" packet, called "peer leave" in the spec, is renamed as the
"peer introduced" packet, and is interpreted as indicating that the connection
is no longer needed as an introducer connection and can be killed if it is
needed for no other reason, rather than that the connection has necessarily
been killed. Older toxcore will continue to kill such connections immediately,
but this doesn't cause any real problem.