don't save groupnumber
So after save and load, the group numbers will be sequential. It's
arguable, but I think on balance this is better than preserving the
group numbers. Note that for the protocol it makes no difference - the
peer connections will go down, and so peers will update the group
numbers they have for us when we reconnect to that in our online packet.