Revert "add TODO on relaying"
This reverts commit aa89ccfebff79f213a48b427ac9db9530ed246ad.
The primitive form of message receipt confirmation afforded by the
current behaviour, in which our own message is echoed only when at least
one peer has received it, is better than nothing. So, we should change
the spec to describe it rather than change the protocol to accord with
the spec.