Refactor peer syncing
We now send each peer in a single packet instead of the entire
peer list in one big packet. This removes a theoretical hard
limit on the number of peers that can join a group and ensures
that the maximum packet size doesn't exceed the MTU of common
network media.

Some unnecessary syncing logic was removed. This will increase
the aggressiveness of syncing with more redundancy, but should
improve overall syncing speed.

Additionally, the ping packet now contains a checksum of the
public key hash of all confirmed peers instead of just the peer
count. This will ensure that sync requests will be made in
the case that all group peers have the same peer count, but a
different set of peers (i.e. group splits).

Finally - an unused 32-bit uint was removed from the sync request
packet.